requirements.txt
ldap3==2.9.1
Django==5.0.6
djangorestframework==3.15.2
django-environ==0.11.2
gunicorn==21.2.0
